+++ apps/block_scout_web/lib/block_scout_web/controllers/transaction_controller.ex
@@ -20,7 +20,10 @@ defmodule BlockScoutWeb.TransactionController do
       :block => :required,
       [created_contract_address: :names] => :optional,
       [from_address: :names] => :optional,
-      [to_address: :names] => :optional
+      [to_address: :names] => :optional,
+      [created_contract_address: :smart_contract] => :optional,
+      [from_address: :smart_contract] => :optional,
+      [to_address: :smart_contract] => :optional
     }
   ]